GHC.InfoProv.Compat

Description

This module provides a stable interface to access GHC's info provenance information. This is helpful for seeing metadata about heap objects.

Documentation

data InfoProv

Constructors

InfoProv

Fields

ipName :: String

ipDesc :: String

ipTyDesc :: String

ipLabel :: String

ipMod :: String

ipSrcFile :: String

ipSrcSpan :: String

whereFrom :: a -> IO (Maybe InfoProv)

Get information about where a value originated from.

This information is stored statically in a binary when super-duper is enabled. The source positions will be greatly improved by also enabled debug information with -g3. Finally you can enable -fdistinct-constructor-tables to get more precise information about data constructor allocations.

The information is collect by looking at the info table address of a specific closure and then consulting a specially generated map (by -finfo-table-map) to find out where we think the best source position to describe that info table arose from.